+++ src/inklet/__init__.py
@@ -49,7 +49,7 @@ from .components import database, feature_matrix, sequence, value_table
 from .render import outline_text, save_pdf, save_svg, to_pdf, to_svg
 from .plot import (
     Panel, Ramp, Scale, axis, band, grouped_band, broken, colorbar, column, dates, facets,
-    histogram, inset, legend, linear, log, panel, ramp, ribbon, row, symlog,
+    histogram, inset, legend, linear, log, panel, ramp, ribbon, row, symlog, upset,
 )
 from .plot import (
     PolarPanel, circular_histogram, circular_mean, polar, theta_ticks,
@@ -467,7 +467,7 @@ __all__ = [
     "panel", "Panel", "row", "column", "axis", "colorbar", "legend",
     "linear", "log", "symlog", "band", "grouped_band", "broken", "dates", "Scale",
     "ramp", "Ramp", "CategorySet", "categories",
-    "inset", "ribbon", "facets", "histogram",
+    "inset", "ribbon", "facets", "histogram", "upset",
     "polar", "PolarPanel", "theta_ticks",
     "circular_mean", "circular_histogram",
     "AnatomyView", "anatomy_view", "model", "solid", "scene", "axes", "cartoon",